Make sure files still exist after 'create_allocation'
Current implementation requires that these files still exist in their
original location so they can be stated and md5summed.  This could
(shourld) be restructured so that it doesn't assume that the original
files are still there.  Another possibility is to do the stating and
such before the files are moved to the allocation.